Gate a curated warning set as errors
Turn a curated set of compiler warnings into errors for first-party C++
sources, so a known class of latent defects cannot re-enter the tree. The
gate is added after ExternalDependencies and before our subdirectories, so
dependency builds are never affected; each flag is its own generator
expression because a semicolon-list inside one genex is split by
add_compile_options.

Gated classes: old-style-cast, unused-variable, unused-const-variable,
shadow, deprecated-volatile, implicit-int-conversion, shorten-64-to-32,
double-promotion.

Clear all 128 pre-existing sites across 42 files. Old-style casts become
the correct C++ cast; unused variables get maybe_unused rather than deletion
to preserve intentional test scaffolding; shadows are renamed. Numeric
narrowings get an explicit cast only where the value is provably in range
(byte assembly into a same-width integer, masked fields, bounded indices);
INFINITY becomes numeric_limits double infinity.

Harden the five genuinely-unbounded CLI parses in the executive instead of
silencing them: rt-max-lag, startup-delay, shutdown-after, watchdog-interval
and profile-interval took a 64-bit stoul straight into a uint32 field with
no bound check. A new parseClampedU32 saturates at UINT32_MAX and each site
logs a new WARN_CLI_ARG_CLAMPED warning rather than wrapping silently.

Verified: keep-going build reports zero -Werror errors (was 142); full unit
suite green after updating the EOE_EXECUTIVE marker test and toString for
the new warning code.

+++ src/system/core/components/action/apex/utst/ResourceCatalog_uTest.cpp
@@ -81,7 +81,9 @@ TEST(ResourceCatalog, WatchpointCatalogDuplicate) {
 TEST(ResourceCatalog, WatchpointCatalogSortedLookup) {
   WatchpointCatalog cat;
 
-  for (std::uint16_t id : {50, 10, 30, 20, 40}) {
+  for (std::uint16_t id : {static_cast<std::uint16_t>(50), static_cast<std::uint16_t>(10),
+                           static_cast<std::uint16_t>(30), static_cast<std::uint16_t>(20),
+                           static_cast<std::uint16_t>(40)}) {
     WatchpointDef def{};
     def.watchpointId = id;
     def.eventId = id;
@@ -89,7 +91,9 @@ TEST(ResourceCatalog, WatchpointCatalogSortedLookup) {
   }
 
   EXPECT_EQ(cat.size(), 5U);
-  for (std::uint16_t id : {10, 20, 30, 40, 50}) {
+  for (std::uint16_t id : {static_cast<std::uint16_t>(10), static_cast<std::uint16_t>(20),
+                           static_cast<std::uint16_t>(30), static_cast<std::uint16_t>(40),
+                           static_cast<std::uint16_t>(50)}) {
     const auto* found = cat.findById(id);
     ASSERT_NE(found, nullptr);
     EXPECT_EQ(found->eventId, id);
